Update .travis.yml for Step 2
The team would like to know that every change they make will be validated with these test cases. Testing should only occur after the jar files are built. Disable any command step, such as the “coming soon” command that was previously causing the testing to fail when the command did not exist.