Update .travis.yml for Step 3
Your team lead is concerned about the amount of code changes they’ve been introducing and is worried that the team is not keeping up with test cases needed for their code changes.
The lead also provided a quick bash command that can show the number of lines covered and not covered from the reports.