Update .travis.yml for Step 5
Set up Travis-CI configuration to run this additional step before any deployment step. Ideally, if the step fails, we should have an "errored" build so that deployment does not occur. We don’t have a deployment yet, so for now just stub a deployment so that we can run the validation on any branch.