Update .travis.yml for Step 9
Add the secure credential to the Travis-CI configuration and validate that the docker login happens as a part of deployment. Use a script provider for Travis-CI to manage this deployment. And make sure to include all deployment commands as a single line for the script configuration, as this provider does not support multi-step script configurations. Make sure to enable docker services on your travis.yml configuration.